% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if_else.R
\name{do_if_else}
\alias{do_if_else}
\title{Conditionally apply expressions on a data object}
\usage{
do_if_else(.data, condition, .if, .else = identity)
}
\arguments{
\item{.data}{Input data}

\item{condition}{A logical value to determine whether to use .if or .else}

\item{.if}{Formula or function to apply to intput data when condition is TRUE}

\item{.else}{Formula or function to apply to intput data when condition is FALSE}
}
\value{
Output of appropriate .if/.else call
}
\description{
Conditionally apply expressions on a data object
}
