% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{untruth_auth}
\alias{untruth_auth}
\title{Authenticate with Truth Social}
\usage{
untruth_auth(
  CLIENT_ID = Sys.getenv("truth_social_client_id"),
  CLIENT_SECRET = Sys.getenv("truth_social_client_secret"),
  username = Sys.getenv("truth_social_user"),
  password = Sys.getenv("truth_social_pw")
)
}
\arguments{
\item{CLIENT_ID}{The client ID provided by Truth Social.}

\item{CLIENT_SECRET}{The client secret provided by Truth Social.}

\item{username}{The username to use for authentication.}

\item{password}{The password to use for authentication.}
}
\value{
A character string containing the authentication token.
}
\description{
This function authenticates with Truth Social using the provided client ID, client secret, username, and password. If a valid authentication token is already stored in the \code{truth_social_token} environment variable, the function will ask if the user wants to refresh the token.
}
