% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{untruth_lookup_users}
\alias{untruth_lookup_users}
\title{Look up a user by their handle}
\usage{
untruth_lookup_users(user_handle, monitor = F, switch_token = NULL, ...)
}
\arguments{
\item{user_handle}{The user's handle.}

\item{monitor}{Whether to monitor the rate limit.}

\item{switch_token}{Whether to switch the token.}

\item{...}{Additional parameters to pass to the \code{httr::GET} function.}
}
\value{
A data frame containing information about the user.
}
\description{
This function allows you to look up a user by their handle.
}
\examples{
untruth_lookup_users("realdonaldtrump")
}
