% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{untruth_search}
\alias{untruth_search}
\title{Search for content on Truth Social}
\usage{
untruth_search(
  what_are_you_looking_for,
  search_type = "statuses",
  limit = 40,
  size = 80,
  verbose = T,
  retry_on_timeout = T,
  switch_token = NULL,
  monitor = F,
  ...
)
}
\arguments{
\item{what_are_you_looking_for}{A character string specifying the search query.}

\item{search_type}{A character string specifying the type of content to search for. Valid values are "statuses" and "hashtags".}

\item{limit}{An integer specifying the maximum number of results to return.}

\item{size}{An integer specifying the size of the results pages to return.}

\item{verbose}{A logical value indicating whether to print status messages to the console.}

\item{retry_on_timeout}{A logical value indicating whether to retry the search if a timeout occurs.}

\item{switch_token}{A logical value indicating whether to use a randomly selected token from the available tokens \verb{DON'T USE. This is EXPERIMENTAL}.}

\item{monitor}{A logical value indicating whether to monitor the rate limit status.}

\item{...}{Additional arguments to pass to the httr::GET() function.}
}
\value{
A data frame containing the search results.
}
\description{
This function searches for content on Truth Social using the provided query string and search type. The number of results to return is determined by the \code{limit} argument, and the size of the results pages is determined by the \code{size} argument. If the \code{switch_token} argument is provided, the function will use a randomly selected token from the available tokens.
}
