% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{untruth_trends}
\alias{untruth_trends}
\title{Get trending hashtags or truths from Truth Social}
\usage{
untruth_trends(type = "hashtags", switch_token = NULL)
}
\arguments{
\item{type}{A character string specifying the type of content to retrieve. Valid values are "hashtags" and "truths".}

\item{switch_token}{A logical value indicating whether to use a randomly selected token from the available tokens \verb{DON'T USE. This is EXPERIMENTAL}.}
}
\value{
A data frame containing the trending hashtags or truths, as well as a timestamp indicating when the data was retrieved.
}
\description{
This function retrieves trending hashtags or truths from Truth Social. The type of content to retrieve is determined by the \code{type} argument. If the \code{switch_token} argument is provided, the function will use a randomly selected token from the available tokens.
}
