% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{untruth_user_statuses}
\alias{untruth_user_statuses}
\title{Retrieve user statuses on Truth Social}
\usage{
untruth_user_statuses(
  user_handle = NULL,
  account_id = NULL,
  limit = 40,
  size = 120,
  verbose = T,
  retry_on_timeout = T,
  switch_token = NULL,
  monitor = F,
  ...
)
}
\arguments{
\item{user_handle}{The user's handle on Truth Social}

\item{account_id}{The user's account ID on Truth Social}

\item{limit}{The number of statuses to retrieve}

\item{size}{The number of statuses per pagination}

\item{verbose}{Whether to print information about pagination}

\item{retry_on_timeout}{Whether to retry the request if a timeout occurs}

\item{switch_token}{Whether to use a different API token for this request}

\item{monitor}{Whether to monitor the API rate limit}
}
\value{
A data frame containing the statuses
}
\description{
Retrieve user statuses on Truth Social
}
\examples{
untruth_user_statuses("truth_social", account_id = "107780257626128497")
untruth_user_statuses("truth_social")
}
